backend/app/requirements.txt
Flask
mysql-connector-python==9.0.0
